refactor: extract policy cli module
